requirements.txt
google-api-python-client==1.7.2
google-auth-oauthlib==0.4.1
yt-dlp==2023.3.4
requests==2.28.2
python-dotenv==1.0.0